layoutWeight支持float类型

+++ frameworks/bridge/declarative_frontend/jsview/models/view_abstract_model_impl.cpp
@@ -561,7 +561,7 @@ void ViewAbstractModelImpl::SetLayoutPriority(int32_t priority)
 
 void ViewAbstractModelImpl::SetPixelRound(uint8_t value) {}
 
-void ViewAbstractModelImpl::SetLayoutWeight(int32_t value)
+void ViewAbstractModelImpl::SetLayoutWeight(float value)
 {
     auto flex = ViewStackProcessor::GetInstance()->GetFlexItemComponent();
     flex->SetFlexWeight(value);
